refactor: 컴포넌트 및 타입 이름 변경
